Split ApparentHorizons/Tags.hpp
This is part of #3561.

  1. NumericalAlgorithms/SphericalHarmonics/Tags.hpp gets all the tags
     that used to be in ApparentHorizons/Tags.hpp and that are
     associated with StrahlkorperFunctions.

  2. ApparentHorizons/Tags.hpp still has StrahlkorperGr tags and
     tags for the AH finder.  Those will be split again in a later
     PR that moves StrahlkorperGr tags elsewhere.

Note:
 * NumericalAlgorithms/SphericalHarmonics/TagsDeclarations.hpp
   and ApparentHorizons/TagsDeclarations.hpp were split in the same way
   as the corresponding Tags.hpp files.

 * Test_Tags.cpp has also been split in the same way.

 * ApparentHorizons/TagsTypeAliases.hpp has been moved to
   NumericalAlgorithms/SphericalHarmonics.
