+++ src/ApparentHorizons/TagsDeclarations.hpp
@@ -4,33 +4,6 @@
 #pragma once
 
 namespace StrahlkorperTags {
-
-template <typename Frame>
-struct Strahlkorper;
-template <typename Frame>
-struct ThetaPhi;
-template <typename Frame>
-struct Rhat;
-template <typename Frame>
-struct Jacobian;
-template <typename Frame>
-struct InvJacobian;
-template <typename Frame>
-struct InvHessian;
-template <typename Frame>
-struct Radius;
-template <typename Frame>
-struct CartesianCoords;
-template <typename Frame>
-struct DxRadius;
-template <typename Frame>
-struct D2xRadius;
-template <typename Frame>
-struct LaplacianRadius;
-template <typename Frame>
-struct NormalOneForm;
-template <typename Frame>
-struct Tangents;
 template <typename Frame>
 struct EuclideanAreaElement;
 template <typename IntegrandTag, typename Frame>
@@ -65,11 +38,9 @@ template <typename Frame>
 struct DimensionfulSpinVector;
 template <typename Frame>
 struct DimensionfulSpinVectorCompute;
-
 }  // namespace StrahlkorperTags
 
-namespace StrahlkorperGr {
-namespace Tags {
+namespace StrahlkorperGr::Tags {
 template <typename Frame>
 struct AreaElement;
 template <typename IntegrandTag, typename Frame>
@@ -93,6 +64,4 @@ template <typename Frame>
 struct DimensionlessSpinMagnitude;
 template <typename Frame>
 struct DimensionlessSpinMagnitudeCompute;
-
-}  // namespace Tags
-}  // namespace StrahlkorperGr
+}  // namespace StrahlkorperGr::Tags
