+++ tests/Unit/Helpers/NumericalAlgorithms/SphericalHarmonics/StrahlkorperTestHelpers.hpp
