+++ tests/Unit/ParallelAlgorithms/Interpolation/Test_ObserveTimeSeriesOnSurface.cpp
@@ -12,6 +12,7 @@
 #include <utility>
 #include <vector>
 
+#include "ApparentHorizons/Tags.hpp"
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/DataBox/Tag.hpp"
 #include "DataStructures/DataVector.hpp"
